+++ lib/src/aes_cbc_cipher_base.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:aes_cbc_cipher/src/index.dart';
 import 'package:encrypt/encrypt.dart' as enc;
 
@@ -156,6 +158,100 @@ class AESCBCCipher {
     return strToReturn;
   }
 
+  /// Encrypts a file located at the given [filePath] using AES-CBC cipher.
+  /// Returns the path of the encrypted file.
+  ///
+  /// Throws an assertion error if the [enc.Encrypter] is not initialized.
+  /// The [enc.Encrypter] must be initialized by calling the [init] method of the AESCBCCipher class.
+  ///
+  /// The file is read as bytes and encrypted using the [enc.Encrypter] with the [enc.IV] as the initialization vector.
+  /// The encrypted bytes are then written to a new file with the same name as the original file, but with the extension ".aes".
+  /// The encrypted file is saved in the application documents directory.
+  Future<String?> encryptFile({
+    required String filePath,
+  }) async {
+    assert(
+      _encrypter != null,
+      'init method not called yet. Please initialize the AESCBCCipher first.',
+    );
+    final File file = File(filePath);
+    final List<int> bytesToEncrypt = await file.readAsBytes();
+
+    final enc.Encrypted encrypted = _encrypter!.encryptBytes(
+      bytesToEncrypt,
+      iv: _iv,
+    );
+
+    final String fileName = filePath.split('/').last.split('.').first;
+
+    final String dir = filePath
+        .split('/')
+        .sublist(0, filePath.split('/').length - 1)
+        .join('/');
+
+    final File encryptedFile = File('$dir/$fileName.aes');
+    await encryptedFile.writeAsBytes(encrypted.bytes);
+
+    return encryptedFile.path;
+  }
+
+  /// Decrypts a file using AES CBC cipher.
+  ///
+  /// This method takes in the path of an encrypted file and decrypts it using the AES CBC cipher algorithm.
+  /// The decrypted file is saved in the same directory as the encrypted file with the correct file extension.
+  ///
+  /// Parameters:
+  /// - `encryptedFilePath`: The path of the encrypted file to be decrypted.
+  ///
+  /// Returns:
+  /// A [Future] that completes with the path of the decrypted file.
+  ///
+  /// Throws:
+  /// - An assertion error if the `init` method has not been called to initialize the AES CBC cipher.
+  Future<String?> decryptFile({
+    required String encryptedFilePath,
+  }) async {
+    assert(
+      _encrypter != null,
+      'init method not called yet. Please initialize the AESCBCCipher first.',
+    );
+
+    final File encryptedFile = File(encryptedFilePath);
+    final List<int> encryptedBytes = await encryptedFile.readAsBytes();
+
+    final enc.Encrypted encrypted = enc.Encrypted(
+      Uint8List.fromList(encryptedBytes),
+    );
+
+    final List<int> decryptedBytes = _encrypter!.decryptBytes(
+      encrypted,
+      iv: _iv,
+    );
+
+    final String base64Decoded = base64Encode(
+      decryptedBytes,
+    );
+
+    // Find the directory where the encrypted file is stored
+    final encryptedDirectory = encryptedFile.parent;
+
+    // Infer the file type from the base64 string
+    final fileType = _inferFileTypeFromBase64(base64Decoded);
+
+    // Get the file name from the encrypted file path
+    final String fileName = encryptedFilePath.split('/').last.split('.').first;
+
+    // Construct the new file path with the correct file extension
+    final decryptedFilePath =
+        '${encryptedDirectory.path}/$fileName.$fileType';
+
+    // Write the decrypted file to the new path
+    final decryptedFile = File(decryptedFilePath);
+    await decryptedFile.writeAsBytes(decryptedBytes);
+
+    return decryptedFilePath;
+  }
+
   //* MARK: - Private Functions
   //? =========================================================
   /// Private function to test the initialized values
@@ -179,6 +275,23 @@ class AESCBCCipher {
         'AESCBCCipher initialization failed. Please check the key and iv values.',
       );
       return e;
+    }
+  }
+
+  // Function to infer file type from base64 string
+  String _inferFileTypeFromBase64(String base64String) {
+    final String firstChar = base64String.substring(0, 1);
+    switch (firstChar) {
+      case 'i':
+        return 'png';
+      case 'R':
+        return 'gif';
+      case '/':
+        return 'jpg';
+      case 'J':
+        return 'pdf';
+      default:
+        return 'png';
     }
   }
 }