+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.2+1
+
+- Improved file encryption functionality: Encrypted files will now be stored in the same directory as the original files.
+- Enhanced file decryption feature: Decrypted files will be stored in the directory where the corresponding encrypted files are located.
+
+## 1.0.2
+
+- Added 2 more functions named `encryptFile` & `decryptFile`.
+
 ## 1.0.1
 
 - Added a new way of encryption with resembles to the [CryptoJS] package for Node.JS.